requirements.txt
acme==0.22.2
asn1crypto==0.24.0
certbot==0.22.2
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
colorclass==2.2.0
ConfigArgParse==0.13.0
configobj==5.0.6
cryptography==2.2.1
dj-static==0.0.6
Django==2.0.3
-e git+https://github.com/yourlabs/django-autocomplete-light.git@28c9ece17118691dbdc97c24c9a0002dfd892517#egg=django_autocomplete_light
django-input-mask==2.0.2
django-localflavor==2.0
django-mysql-pymysql==0.1
django-widget-tweaks==1.4.1
docopt==0.6.2
enum34==1.1.6
funcsigs==1.0.2
future==0.16.0
gevent==1.2.2
greenlet==0.4.13
gunicorn==19.7.1
idna==2.6
ipaddress==1.0.19
josepy==1.0.1
letsencrypt==0.7.0
meld3==1.0.2
mock==2.0.0
mysqlclient==1.3.12
ndg-httpsclient==0.4.4
netifaces==0.10.6
packaging==17.1
parsedatetime==2.4
pbr==3.1.1
pip-upgrader==1.4.4
pkg-resources==0.0.0
psutil==5.4.3
psycopg2==2.7.4
pyasn1==0.4.2
pycparser==2.18
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyRFC3339==1.0
pytz==2018.3
requests==2.18.4
six==1.11.0
South==1.0.2
sqlparse==0.2.4
static3==0.7.0
terminaltables==3.1.0
urllib3==1.22
virtualenv==15.2.0
zope.component==4.4.1
zope.event==4.3.0
zope.hookable==4.2.0
zope.interface==4.4.3
